完善框架

+++ src/main/java/cn/hhfarcry/springbootmybatis/example/service/impl/RoleService.java
@@ -1,9 +1,12 @@
 package cn.hhfarcry.springbootmybatis.example.service.impl;
 
+
 import cn.hhfarcry.springbootmybatis.common.base.service.BaseService;
 import cn.hhfarcry.springbootmybatis.common.base.utils.ParamUtils;
+import cn.hhfarcry.springbootmybatis.example.dao.ResourceDao;
 import cn.hhfarcry.springbootmybatis.example.dao.RoleDao;
 import cn.hhfarcry.springbootmybatis.example.dao.RoleResourceDao;
+import cn.hhfarcry.springbootmybatis.example.entity.ResourceEntity;
 import cn.hhfarcry.springbootmybatis.example.entity.RoleEntity;
 import cn.hhfarcry.springbootmybatis.example.entity.RoleResourceEntity;
 import cn.hhfarcry.springbootmybatis.example.service.IRoleService;
@@ -14,10 +17,12 @@ import org.springframework.transaction.annotation.Transactional;
 
 import javax.annotation.PostConstruct;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 /**
- * @program: springbootmybatis
+ * @program: emsog
  * @description: ${description}
  * @author: huanghong
  * @date: 2019-01-16 11:21
@@ -30,6 +35,9 @@ public class RoleService extends BaseService<RoleEntity> implements IRoleService
     private RoleDao roleDao;
 
     @Autowired
+    private ResourceDao resourceDao;
+
+    @Autowired
     private RoleResourceDao roleResourceDao;
 
     @Override
@@ -40,6 +48,15 @@ public class RoleService extends BaseService<RoleEntity> implements IRoleService
 
     @Override
     public String insertRole(RoleEntity param) {
+        Map<String,Object>map = new HashMap<>();
+        map.put("roleName",param.getRoleName());
+        List<RoleEntity>existRolename = roleDao.selectByEntity(map);
+        if(CollectionUtils.isNotEmpty(existRolename)){
+            if((ParamUtils.isNotBlank(param.getId()) && !existRolename.get(0).getId().equals(param.getId()))
+                    || (ParamUtils.isBlank(param.getId()))){
+                return "角色名已存在";
+            }
+        }
         super.buildUser(param);
         roleDao.insertOrUpdateByEntity(param);
         return "ok";
@@ -47,12 +64,20 @@ public class RoleService extends BaseService<RoleEntity> implements IRoleService
 
     @Override
     public String bindroleResources(String roleId, List<String> resourceIds) {
+        RoleEntity existRole = roleDao.getOne(ParamUtils.strTIntger(roleId));
+        if(ParamUtils.isBlank(existRole)){
+            return "角色不存在";
+        }
+
+        //删除角色现有的资源
         roleResourceDao.deleteRoleResources(ParamUtils.strTIntger(roleId));
+        //查询resourceIds中有效的id
+        List<ResourceEntity>existRes = resourceDao.getResourcesByResourceIds(ParamUtils.strlisttointlist(resourceIds));
         List<RoleResourceEntity>params = new ArrayList<>();
-        for (String resourceId : resourceIds) {
+        for (ResourceEntity resourceEntity : existRes) {
             RoleResourceEntity param = new RoleResourceEntity();
             param.setRoleId(ParamUtils.strTIntger(roleId));
-            param.setResourceId(ParamUtils.strTIntger(resourceId));
+            param.setResourceId(resourceEntity.getId());
             params.add(param);
         }
         if(CollectionUtils.isNotEmpty(params)){
@@ -62,7 +87,9 @@ public class RoleService extends BaseService<RoleEntity> implements IRoleService
     }
 
     @Override
-    public List<RoleEntity> getRolesByUserId(Integer userId) {
-        return roleDao.getRolesByUserId(userId);
+    public List<RoleEntity> getRolesByUserId(String userId) {
+        return roleDao.getRolesByUserId(ParamUtils.strTIntger(userId));
     }
+
+
 }
